+++ src/scanner/RepositoryConfig.ts
@@ -1,0 +1,6 @@
+export interface RepositoryConfig {
+  remoteUrl: string | undefined;
+  host: string | undefined;
+  protocol: string | undefined;
+  baseUrl: string | undefined;
+}