+++ src/scanner/Scanner.ts
@@ -1,41 +1,41 @@
+import { cli } from 'cli-ux';
 import debug from 'debug';
 import fs from 'fs';
 import { inject, injectable, multiInject } from 'inversify';
+import _ from 'lodash';
 import os from 'os';
 import path from 'path';
 import git from 'simple-git/promise';
 import url from 'url';
 import { inspect } from 'util';
+import { ArgumentsProvider } from '.';
+import { LanguageContext } from '../contexts/language/LanguageContext';
+import { PracticeContext } from '../contexts/practice/PracticeContext';
+import { ProjectComponentContext } from '../contexts/projectComponent/ProjectComponentContext';
+import { ScannerContext } from '../contexts/scanner/ScannerContext';
+import { AccessType, ScanningStrategy, ServiceType } from '../detectors';
+import { sharedSubpath } from '../detectors/utils';
+import { ErrorFactory } from '../lib/errors';
 import {
   LanguageAtPath,
   PracticeEvaluationResult,
+  PracticeImpact,
   ProjectComponent,
   ProjectComponentFramework,
   ProjectComponentPlatform,
   ProjectComponentType,
-  PracticeImpact,
 } from '../model';
 import { IPracticeWithMetadata } from '../practices/DxPracticeDecorator';
-import { ScannerContextFactory, Types } from '../types';
-import { ScannerUtils } from '../scanner/ScannerUtils';
-import _ from 'lodash';
-import { cli } from 'cli-ux';
-import { ScanningStrategyDetector, ScanningStrategy, ServiceType, AccessType } from '../detectors';
 import { IReporter, PracticeWithContextForReporter } from '../reporters';
+import { ScannerUtils } from '../scanner/ScannerUtils';
 import { FileSystemService } from '../services';
-import { ScannerContext } from '../contexts/scanner/ScannerContext';
-import { sharedSubpath } from '../detectors/utils';
-import { LanguageContext } from '../contexts/language/LanguageContext';
-import { ProjectComponentContext } from '../contexts/projectComponent/ProjectComponentContext';
-import { PracticeContext } from '../contexts/practice/PracticeContext';
-import { ArgumentsProvider } from '.';
-import { ErrorFactory } from '../lib/errors';
+import { DiscoveryContextFactory, Types } from '../types';
+import { ScanningStrategyExplorer } from './ScanningStrategyExplorer';
 
 @injectable()
 export class Scanner {
-  private readonly scanStrategyDetector: ScanningStrategyDetector;
-  private readonly scannerContextFactory: ScannerContextFactory;
-  private readonly reporters: IReporter[];
+  private readonly scanStrategyExplorer: ScanningStrategyExplorer;
+  private readonly discoveryContextFactory: DiscoveryContextFactory;
   private readonly fileSystemService: FileSystemService;
   private readonly practices: IPracticeWithMetadata[];
   private readonly argumentsProvider: ArgumentsProvider;
@@ -44,44 +44,30 @@ export class Scanner {
   private allDetectedComponents: ProjectComponentAndLangContext[] | undefined;
 
   constructor(
-    @inject(ScanningStrategyDetector) scanStrategyDetector: ScanningStrategyDetector,
-    @inject(Types.ScannerContextFactory) scannerContextFactory: ScannerContextFactory,
-    @multiInject(Types.IReporter) reporters: IReporter[],
+    @inject(ScanningStrategyExplorer) scanStrategyExplorer: ScanningStrategyExplorer,
+    @inject(Types.DiscoveryContextFactory) discoveryContextFactory: DiscoveryContextFactory,
     @inject(FileSystemService) fileSystemService: FileSystemService,
     // inject all practices registered under Types.Practice in inversify config
     @multiInject(Types.Practice) practices: IPracticeWithMetadata[],
     @inject(Types.ArgumentsProvider) argumentsProvider: ArgumentsProvider,
   ) {
-    this.scanStrategyDetector = scanStrategyDetector;
-    this.scannerContextFactory = scannerContextFactory;
-    this.reporters = reporters;
+    this.scanStrategyExplorer = scanStrategyExplorer;
+    this.discoveryContextFactory = discoveryContextFactory;
     this.fileSystemService = fileSystemService;
     this.practices = practices;
     this.argumentsProvider = argumentsProvider;
+
     this.d = debug('scanner');
     this.allDetectedComponents = undefined;
   }
 
-  async fix(practicesWithContext: PracticeWithContext[]) {
-    if (!this.argumentsProvider.fix) return;
-    const fixablePractice = (p: PracticeWithContext) => p.practice.fix && p.evaluation === PracticeEvaluationResult.notPracticing;
-    const fixPatternMatcher = this.argumentsProvider.fixPattern ? new RegExp(this.argumentsProvider.fixPattern, 'i') : null;
-    const shouldFix = (p: PracticeWithContext) => {
-      const fixFromCli = fixPatternMatcher ? fixPatternMatcher.test(p.practice.getMetadata().id) : undefined;
-      const practiceConfig = p.componentContext.configProvider.getOverriddenPractice(p.practice.getMetadata().id);
-      const fixFromConfig = practiceConfig?.fix;
-      return fixFromCli !== undefined ? fixFromCli : fixFromConfig !== undefined ? fixFromConfig : true;
-    };
-    await Promise.all(
-      practicesWithContext
-        .filter(fixablePractice)
-        .filter(shouldFix)
-        .map((p) => p.practice.fix!(p.practiceContext)),
-    );
-  }
-
   async scan({ determineRemote } = { determineRemote: true }): Promise<ScanResult> {
-    let scanStrategy = await this.scanStrategyDetector.detect();
+    const repositoryConfig = await this.scanStrategyExplorer.explore();
+    this.d(`Repository Config: ${inspect(repositoryConfig)}`);
+
+    const discoveryContext = this.discoveryContextFactory(repositoryConfig);
+
+    let scanStrategy = await discoveryContext.scanningStrategyDetector.detect();
     this.d(`Scan strategy: ${inspect(scanStrategy)}`);
     if (determineRemote && (scanStrategy.serviceType === undefined || scanStrategy.accessType === AccessType.unknown)) {
       return {
@@ -93,7 +79,7 @@ export class Scanner {
     }
     scanStrategy = await this.preprocessData(scanStrategy);
     this.d(`Scan strategy (after preprocessing): ${inspect(scanStrategy)}`);
-    const scannerContext = this.scannerContextFactory(scanStrategy);
+    const scannerContext = discoveryContext.getScanningContext(scanStrategy);
     const languagesAtPaths = await this.detectLanguagesAtPaths(scannerContext);
     this.d(`LanguagesAtPaths (${languagesAtPaths.length}):`, inspect(languagesAtPaths));
     const projectComponents = await this.detectProjectComponents(languagesAtPaths, scannerContext, scanStrategy);
@@ -103,11 +89,11 @@ export class Scanner {
 
     let practicesAfterFix: PracticeWithContext[] | undefined;
     if (this.argumentsProvider.fix) {
-      await this.fix(practicesWithContext);
+      await this.fix(practicesWithContext, scanStrategy);
       practicesAfterFix = await this.detectPractices(projectComponents);
     }
 
-    await this.report(practicesWithContext, practicesAfterFix);
+    await this.report(scannerContext.reporters, practicesWithContext, practicesAfterFix);
     this.d(
       `Overall scan stats. LanguagesAtPaths: ${inspect(languagesAtPaths.length)}; Components: ${inspect(
         this.allDetectedComponents!.length,
@@ -136,6 +122,31 @@ export class Scanner {
     }
 
     cli.action.stop();
+  }
+
+  async fix(practicesWithContext: PracticeWithContext[], scanningStrategy?: ScanningStrategy) {
+    if (!this.argumentsProvider.fix) return;
+    const fixablePractice = (p: PracticeWithContext) => p.practice.fix && p.evaluation === PracticeEvaluationResult.notPracticing;
+    const fixPatternMatcher = this.argumentsProvider.fixPattern ? new RegExp(this.argumentsProvider.fixPattern, 'i') : null;
+    const shouldFix = (p: PracticeWithContext) => {
+      const fixFromCli = fixPatternMatcher ? fixPatternMatcher.test(p.practice.getMetadata().id) : undefined;
+      const practiceConfig = p.componentContext.configProvider.getOverriddenPractice(p.practice.getMetadata().id);
+      const fixFromConfig = practiceConfig?.fix;
+      return fixFromCli !== undefined ? fixFromCli : fixFromConfig !== undefined ? fixFromConfig : true;
+    };
+    await Promise.all(
+      practicesWithContext
+        .filter(fixablePractice)
+        .filter(shouldFix)
+        .map((p) =>
+          p.practice.fix!({
+            ...p.practiceContext,
+            scanningStrategy,
+            config: p.componentContext.configProvider.getOverriddenPractice(p.practice.getMetadata().id),
+            argumentsProvider: this.argumentsProvider,
+          }),
+        ),
+    );
   }
 
   /**
@@ -243,7 +254,11 @@ export class Scanner {
   /**
    * Report result with specific reporter
    */
-  private async report(practicesWithContext: PracticeWithContext[], practicesWithContextAfterFix?: PracticeWithContext[]): Promise<void> {
+  private async report(
+    reporters: IReporter[],
+    practicesWithContext: PracticeWithContext[],
+    practicesWithContextAfterFix?: PracticeWithContext[],
+  ): Promise<void> {
     const pwcForReporter = (p: PracticeWithContext) => {
       const config = p.componentContext.configProvider.getOverriddenPractice(p.practice.getMetadata().id);
       const overridenImpact = config?.impact;
@@ -259,16 +274,16 @@ export class Scanner {
     };
     const relevantPractices: PracticeWithContextForReporter[] = practicesWithContext.map(pwcForReporter);
 
-    this.d(`Reporters length: ${this.reporters.length}`);
+    this.d(`Reporters length: ${reporters.length}`);
     if (this.argumentsProvider.fix) {
       const relevantPracticesAfterFix: PracticeWithContextForReporter[] = practicesWithContextAfterFix!.map(pwcForReporter);
       await Promise.all(
-        this.reporters.map(async (r) => {
+        reporters.map(async (r) => {
           this.argumentsProvider.fix ? await r.report(relevantPractices, relevantPracticesAfterFix) : await r.report(relevantPractices);
         }),
       );
     } else {
-      await Promise.all(this.reporters.map(async (r) => await r.report(relevantPractices)));
+      await Promise.all(reporters.map(async (r) => await r.report(relevantPractices)));
     }
 
     if (this.allDetectedComponents!.length > 1 && !this.argumentsProvider.recursive) {